requirements.txt
aiogram==2.25.2
uvloop
ujson
aiohttp[speedups]

# Other packages
python-dotenv
characterai
